refactor(pkg): move PackageFilter to separate file
